Rename query file.
